ndisasm: add -z option for data size; clean up address handling
- Add a -z option to set the data size to be read (useful with -e)
- Clean up at least some remaining 32-bit assumptions.

+++ disasm/sync.h
@@ -9,7 +9,7 @@
 #define NASM_SYNC_H
 
 void init_sync(void);
-void add_sync(uint64_t position, uint32_t length);
-uint64_t next_sync(uint64_t position, uint32_t *length);
+void add_sync(uint64_t position, uint64_t length);
+uint64_t next_sync(uint64_t position, uint64_t *length);
 
 #endif
